+++ serde_with_macros/tests/compile-fail/skip-none-not-struct.rs
@@ -1,5 +1,3 @@
-extern crate serde_with_macros;
-
 use serde_with_macros::skip_serializing_none;
 
 #[skip_serializing_none]