+++ serde_with_macros/tests/version_numbers.rs
@@ -1,5 +1,3 @@
-extern crate version_sync;
-
 #[test]
 fn test_html_root_url() {
     version_sync::assert_html_root_url_updated!("src/lib.rs");