+++ tests/version_numbers.rs
@@ -1,5 +1,3 @@
-extern crate version_sync;
-
 use version_sync::{
     assert_contains_regex, assert_html_root_url_updated, assert_markdown_deps_updated,
 };