+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+* Convert the code to use 2018 edition.
+
 ## [1.4.0]
 
 ### Added
@@ -18,7 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 * Bump minimal Rust version to 1.36.0
-    * Support Rust Edition 2018
+    * Supports Rust Edition 2018
     * version-sync depends on smallvec which requires 1.36
 * Improved CI pipeline by running `cargo audit` and `tarpaulin` in all configurations now.